+++ src/index.ts
@@ -8,6 +8,8 @@ export * from "./local-storage-provider/local-storage-provider-context";
 export * from "./local-storage-provider/local-storage-provider";
 export * from "./popup-menu/popup-menu";
 export * from "./table-view/table-view";
+export * from "./pagination/pagination";
+export * from "./busy-button/busy-button";
 
 export * from "./form/field-decoration";
 export * from "./form/custom-field";
@@ -20,10 +22,9 @@ export * from "./form/file-field";
 export * from "./form/date-picker";
 export * from "./form/date-field";
 export * from "./form/time-picker";
-export * from "./form/time-field";
 
+export * from "./form/time-field";
 export * from "./widgets/loader";
-export * from "./widgets/busy-button";
 export * from "./widgets/confirm-dialog";
 export * from "./widgets/breadcrumbs";
 export * from "./widgets/page-actions";